Fix a few mutt_extract_token() callers dest parameters.
Some callers performed a mutt_buffer_init(), called
mutt_extract_token() and then assumed that the buffer->data would
never be NULL.  Prior to the "fix" in commit f420be68, this was not
true, even if MoreArgs() returns true, because the argument could be
just ''.

The parse_keymap() change fixed a possible segv.  Don't use the buffer
pool, because it returns buf->data to the caller.

I don't believe the other two would segv, but they did still rely on
the old behavior where the the buffer would have a '\0' appended
always.  So, change them to use the buffer pool to be cleaner.