Parse headers on demand for the list menu.
This will be a bit slower over IMAP from the index, but it removes
storing all the new headers in the header cache.  Since those are not
used for anything except the list menu, and the list menu is
infrequently used, it's a bit wasteful to store them in the header
cache.