Add "headers" parameter to mx_open_message().
This will allow some operations to retrieve only headers, such as ~h
pattern matching or the new list menu.

Modify the IMAP and POP3 implementation to retrieve only headers when
the parameter is set.

Headers-only will use the message cache if one exists, but will not
populate the message cache (since the body of the message is not
downloaded.)