Create $ssl_verify_host_override.
This allows manually specifying the host name to verify a server
certificate against.

I'm not sure I like the idea, but since Mutt already allows turning
hostname verification off, it's not any worse to leave it on but
manually set the expected hostname in the certificate.

Thanks to eudyptes-pachyrynchus for the merge request, which prompted
me to take another look at the ticket.