Fix offset to use LOFF_T in a couple places.
The BODY->hdr_offset was incorrectly of type long, which could result
in corrupted >2gb mbox files in some circumstances.

The uses in mh_rewrite_message() are not as serious, but they should
still be of type LOFF_T.

I found both of these fixes in a patch file,
bug-676388-largefile.patch, in the openSUSE mutt src rpm.  It looks
like Harald Koenig was the original author of a larger patch in that
openSUSE ticket, which was reduced over time as fixes were made to
Mutt.

Note that patch file also incorrectly adjusted old_hdr_lines in mh.c.
I've removed that part, as HEADER->lines is type int.

Unfortunately, the BODY->hdr_offset type change will result in a
header cache change (i.e. invalidation).  I'm not enthused about doing
that during a stable release, but the change is important enough to
merit it.