Correct length to use LOFF_T.
This is a part two, made in master, to the stable branch commit
11b18027.  These are mostly length adjustments to use LOFF_T, matching the
BODY->length type.

An argument could be made for size_t instead, and a few places in Mutt
do assign between those types.  I've used LOFF_T because off_t is a
signed integer.  Some changes in this commit affect loops that
decrement a length pointer while > 0.  Switching to a size_t could
create a wraparound infinite loop bug.

This also changes the Content-Length header parser to use atoll()
intead of atol().  I noticed from the man page that atol() doesn't
seem to return -1 on error.  But I've kept the check anyway.