Fix header cache BDB version checking.
Update BDB versions up to 6.2.

The scheme used here is, as Oswald nicely puts it, "totally insane",
and seems to come straight from the initial header cache commit in
c11667eaa3.

For a stable release, I'm just going to add a few more Jenga pieces to
the top.

However, going forward this is unmaintainable.  Oswald reports using a
simple check for "db.h" and "-ldb" for about two decades with no
issues.  I will make that change in master later this week.