Simplify BDB configure.ac check.
Remove the long list of directory and library names to scan, and just
look for -ldb and db.h.  Use $with_bdb to search an alternative
directory, just as the other header cache library searches do.

If someone screams, we can enhance a *little* bit, but I think this is
sufficient, and more important, maintainable.