Update header cache checking order and --with options.
Automatically scan in the order: kyotocabinet, tokyocabinet, lmdb,
qdbm, gdbm, bdb.

Turn all the backend options into '--with' options and add a dir
parameter.  Specifying a single --with option declares the desired
backend, turning off scanning.  Specifying --without skips scanning
for a backend.

Update the INSTALL file to make that clear.  This actually wasn't even
clear to me until I studied it now, and for years I've been using an
obfuscated combination of --without and --with to specify the backend,
when I just needed to use --with.