Add configure checks for gssapi library and functions.
rkta on irc reported a build failure because the gss headers were not
present.  Add explicit checks for the required headers.  Also add a
check for gss_init_sec_context() when using the output of krb-config
to configure flags.

Abort configure if the library/header checks don't pass.  It would
previously continue on, setting USE_GSS and need_gss, which would just
result in a build error.