Add explicit flag to check for postponed message on <mail>.
This used to check for:
  !flags && !msg
which was still relatively clear that it was a unique situation (just
for <mail> in the index and pager).

With the addition of background editing, it became:
  (sctx->flags == SENDBACKGROUNDEDIT) && !sctx->msg
This is less intuitively understood to be just for the <mail> case,
and thus perhaps easy to create a bug in the future.

Create an explicit flag, SENDCHECKPOSTPONED to perform the check, and
enable for the <mail> operation in the index and pager.