Fix gettext update problems.
The keymap_alldefs.h file was previously generated and scanned inside
po/Makefile.in.in as an additional POTFILES files build dependency and
step.

The file can't be scanned by just adding it to po/POTFILES.in because
it's a generated file that will be in the build directory.  The next
commit will instead add the OPS* files directly to POTFILES.in to be
scanned.

For this commit, remove the generation of keymap_alldefs.h.

Remove the redundant ALL_LINGUAS setting from configure.ac.  Reformat
the LINGUAS file to be on one line, as the documentation shows.  I
think it was okay before, but just to be safe.

Fix L10N scanning by adding '--add-comments=L10N' to XGETTEXT_OPTIONS
in po/Makevars.

Set MSGID_BUGS_ADDRESS to silence a warning.

Add gettext.h to the Makefile.am EXTRA_DIST list.

Remove -Iintl flags from all Makefile.am files, since the intl
directory is no longer bundled in Mutt.