Fix (un)setenv to not return an error with unset env vars.
The code was returning -1 in those cases.  This can cause hook
processors, such as send2-hook, to sleep 1 second without an error
message being set and displayed.  It also causes the hook processor to
abort executing subsequent matching commands.

-1 should be reserved for syntax errors.  For setenv querying, change
the return value to 0.  For unsetenv, generate a message and also
change the return value to 0.

<enter-command> checks if err->data is set to determine whether to
print anything, and the behavior is used in the 'set' commands, so
this is established correct behavior.