Prompt for retry on decrypt failure for mime message attachments.
Closes #331.

Change $forward_decyrpt to a quadoption, defaulting 'yes' for backward
compatibility.  This will allow users to choose on a per-message
attachment basis if desired.

If decoding from $mime_forward_decode fails or decrypting from
$forward_decrypt, notify the user and prompt to retry attaching
without the decode/decrypt.

$forward_decrypt directly invokes decryption and returns -1 on error,
while $mime_forward_decode uses the handler rendering, which returns a
non-fatal 1 (to allow message display of failing encrypted message).
So prompt on either case.

If they decline to retry, or the retry fails, return NULL only on a
hard-fail (-1) return code.  They've been warned and can peek at the
final result.

The send generate_body() code was not checking for a NULL result.
This seems like a bad idea, so add a check and error message in that
case.