Don't install an empty manual.txt file.
Closes #334.

If all the html to txt programs we use (lynx, elinks, w3m) are missing
from the system, the shell redirection willl create an empty file.

Change the install step to use -s "FILE exists and has a size greater
than zero".

Thanks to Mitchell Blank Jr for the ticket and suggested fix.