Fix $ssl_client_cert usage with SMTP.
The ssl and gnutls client-cert setup code was calling
mutt_account_getuser().  This caused two problems.  First, it's not
necessarily the case that there will be a username.  Second,
populating the user would cause smtp_open() to check for AUTH
capabilities and call smtp_auth - even if the user is already
authenticated by the cert.  The server won't advertize AUTH if they
already authenticated, causing a connection abort.

Remove prompt for mutt_account_getuser() in the ssl and gnutls client
certificate connection code.  The SASL code has callbacks, so I don't
understand why it would need this.  Let's take it out and see if
anyone screams 8-P.  If necessary, we can add a mutt_account_getuser()
call to the very beginning of imap_auth_sasl().

Revamp the openssl ssl_passwd_cb() prompt.  From the man pages, it
appears to be used for the cert decryption.  There's no need to call
mutt_account_getuser() and use the generic mutt_account_getpass() just
to read a password in.  Instead create a callback function version to
customize the prompt for a client cert with just the host.

Change the SMTP authentication test to check if the AUTH capabilities
are set, instead of if the user field is set before calling
smtp_auth().