Improve body color matching speed by caching future matches.
On a *very* long body (or header_partial) lines with multiple color
lines that match, performance can be degraded.  For instance if there
were moderately expensive regexp matches against "A" "B" and "C".  A
line with:

  A A A A A B A A C A A A B A A A C

The B and C regexps were scanned again after every "A" match, despite
that the result would be discarded again for the next A match.

Change the color matching to cache the location of each color line
match.  Discard the cache once the match offset passes that point.