Fix body color null_rx handling.
The existing code did not work properly.  If the null_rx occurred
before another matching entry in the list, it would be ignored.  If it
occurred afterward, it could cause the entries to be created
out-of-order which would affect the binary search later on when
rendering.

Instead, use the normal "match assignment" code, but if the best
result is a null_rx match, simply discard it and increment the offset
past the end of the null match.