Fix resolve_types() newline removal and restoration.
If the line had a '\n', the previous code would leave it off.  The
"buf[nl] = '\n'" at the end of the function was overwriting the old
nul character, but there was a new one where the original newline was.

However, if there was not a '\n', the code was removing the nul string
terminator, which could cause very bad things for searches, (which are
the only things that use fmt after resolve_types).  Since this bug has
been there for 12 years, I'm guessing the lines always contain a
trailing '\n'.

Now, there is a potential problem with behavior changes here...  This
could have an effect on the QuoteRegexp or SearchRE matches.  It turns
out SearchRE is compiled with REG_NEWLINE, but the other isn't.  Since
QuoteRegexp is (typically) only looking at the start of a line, I'm
going to take the risk of restoring the '\n'.