Minor fixes to match_body_patterns().
Change the end condition to compare against buflen.  I prefer this
because we are not directly incrementing offset ourselves, but relying
on the rational return values from regexec.  If the regex library
(incorrectly) returned a nul-match at the '\0' terminator for
instance, the new nul-match logic would assign offset just past the
terminator.

Remove the i=0 initialization during definition, because it doesn't
use or need it.