Turn off ~h header-only download when $message_cachedir is set.
The initial pattern match will be slow, but then the body will be in
the message cache.  This is the same as previous behavior, so a
message cache user should not notice any change.

Those not using a message cache will see improvements from the
header-only download, but of course they will have to download the
full headers each time.