Add SigInt handler for pattern functions.
Allow, for example, a ~B against a huge mailbox to be interruped via
ctrl-c.  This commit follows the same behavior as
mutt_search_command() - not prompting and just interrupting.  We could
add a prompt if it's really deemed necessary, but I think if they hit
ctrl-c they want to stop the pattern operation now.

Since it's not practical to roll back, leave the Context in the "in
the middle" state too.  This may actually be desired.