Improve body_pattern and quoting when skipping down.
The previous commit added the MUTT_TYPES flag to the search operation.
However, adding the flag caused body_pattern and classify_quote
matching to occur, as part of searching.

Separate those two to only be done when MUTT_SHOWCOLOR is passed.  Keep
the classify_quote() call in resolve_types(), since otherwise during a
normal display we would end up matching the regexps twice.

Add a flag so the computation only occurs once, instead of each time
the line is shown.

For continuation lines, we need to scan from the start of line.
Otherwise body_patterns and proper quote level coloring won't appear
if we jump down to the middle of a continuation line.

As part of this commit, I noticed is_cont_hdr was not reset when
handling a REDRAW_FLOW.  I believe this also fixes the old trac ticket
issue 3744.