Disable markers when wrap is at column 1.
This can occur when the user resizes the window to a single column, or
if they 'set wrap=1'.

In that case, Mutt was outputting the '+' and setting the initial
output to the second column.  However, if there is only one column,
nothing would then be printed.  When trying to search or jump to end,
Mutt would enter an infinite loop because no forward progress would be
made in the file with each line.

Disable the markers if there is only a single column, so at least one
character of the message will get printed on each line.