Add local_date_header option
Add an option to convert the date used in the Date header into the local
(sender's) timezone. This is the current behavior and the option
defaults to true. Unsetting this option causes the date in the Date
header to be formatted using the GMT timezone.

This option is useful for privacy-sensitive users who may not wish to
divulge their sending timezone.