Trim the ToC for the option list in the manual
Currently, every variable is shown in the table of contents of the
reference manual. Variable list is over 400 items long. This makes the
ToC difficult to work with.

Removing is done by adding an overwriting template for the parent of
variables section (that is sect1 tag with id="variables"). In the
DocBook source matching template calls another template named subtoc
with two parameters: context (self element) and nodes (child
sections). Normally a list of descendant nodes is passed here, but we
don't want that and instead call subtoc with no parameters.

This approach results in semantically correct code being generated in
all modes.

Note, that the ID (variables) will have to be kept in sync with XML
generating code, but that is already hardcoded elsewhere.