fix gcc-11 build warnings for H_TO_INTL()
The macro H_TO_INTL() in mutt_idna.c, causes a lot of build warnings
about the indentation of the if () line to try to warn the developer
that it looks like this might be a bug:

mutt_idna.c: In function ‘mutt_env_to_intl’:
mutt_idna.c:327:5: warning: this ‘if’ clause does not guard... [-Wmisleading-indentation]
  327 |     if (tag) *tag = #a; e = 1; err = NULL;              \
      |     ^~
mutt_idna.c:333:3: note: in expansion of macro ‘H_TO_INTL’
  333 |   H_TO_INTL(return_path);
      |   ^~~~~~~~~
mutt_idna.c:327:25: note: ...this statement, but the latter is misleadingly indented as if it were guarded by the ‘if’
  327 |     if (tag) *tag = #a; e = 1; err = NULL;              \
      |                         ^
mutt_idna.c:333:3: note: in expansion of macro ‘H_TO_INTL’
  333 |   H_TO_INTL(return_path);
      |   ^~~~~~~~~

This goes on for many lines.

Fix this up by properly indenting the lines so that gcc "knows" we mean
well here.