Do not set From reverse name on postponed or to be resend messages
Messages to be resend or resumed from postponed already had
reverse names and alternates evaluated the first time they entered
the compose menu and send-hook were already executed then. The
user may also had edited the From to any other content before
postponing or sending the message. Upon resuming, do not overwrite
with a reverse_name lookup and do not remove the real name if
reverse_realname=no is set.

This restores the old behaviour and fixes the regression
introduced with

    commit 961cf4bed679d9495cc3682d18496a53061d4581
    CommitDate: Mon Feb 1 14:00:56 2021 -0800

        Apply $reverse_name when tag-reply/forwarding too.