Add $pager_skip_quoted_context option.
Using <skip-quoted> the first unquoted line becomes the new top line
displayed in the pager. This leaves the user with no context to know
what the answer refers to.

Add an option to keep some lines of context when using <skip-quoted>.

With $pager_skip_quoted_context set to 5, <skip-quoted> will show at
most 5 lines of the previous quote. If the previous quote is shorter
than 5 lines the whole quote will be displayed.

This option defaults to 0 to remain backwards compatible.