Use memcpy in fseek_last_message() instead of strncpy.
We are copying raw bytes and don't care about embedded nuls, so
eliminate a compiler warning by using memcpy.  BUFSIZ is guaranteed to
be at least 256.

Adjust the weird BUFSIZ + 9 size, which was added by an old commit
86f376dc that appears to have been a dead end.  Also remove the no
longer relevant comment next to the strncpy.

Although it may be "safer" to use strlen() for the mutt_strncmp()
parameter, the code is performing this comparison over and over.  We
already hardcode 7 in the buffer size and memcpy.  So, make this loop
faster by hardcoding it here too.