Improve generic menu type handling.
Some of the menus are directly of MENU_GENERIC type.  In a few places,
Mutt handles those inefficiently or slightly incorrectly.

Remove double "generic" lookups when Mutt is in a generic menu.

Fix the help menu to not show unbound or "generic bindings" for a
generic menu.

For exec completion, always add generic functions to the possible
completion list.