Lazily update header colors after a thread-flag update.
Although there are other issues with threads and ~() color patterns,
this helps with collapsed thread coloring.  The thread-flag update can
change the matching pattern, but that should be determined after all
the thread messages are changed.

For example, a collapsed thread with deleted messages and a color
pattern '~v ~(~D)' should display differently after <undelete-thread>.
Changing it to update colors lazily allows correct coloring (for that
case).