Sort headers by UID when generating a msgset sequence.
When there is a hole in the header cache, the UID numbers are no
longer guaranteed to increase with the index.  This can result in
incorrect msgset values being sent to the server.

Thanks to Pieter-Tjerk de Boer for reporting this problem, with a
detailed description of the problem, and a suggested patch.

Note to backporters:

* this patch depends on the imap_disallow_reopen() added in commit
  6051760c2cf492ada1e06d910c0c2c05607c08bc for release 1.13.4.

* the bug was introduced in release 1.9.0 so there is no need to
  backport earlier than that.