Preserve QRESYNC seqset holes when restoring from header cache.
It's possible for there to be holes in the MSN sequence, and also in
the header cache.  When iterating through the seqset, preserve those
by incrementing msn even if the result isn't found in the header
cache (because it is 0 or because the record is missing in the header
cache).