Improve conststrings.c generation, mainly for GNU Autoconf 2.71.
With GNU Autoconf 2.71, the "ac_cs_config=" line in config.status may
contain sequences like '\'' when a configure argument has spaces in it,
for instance CFLAGS with several options. Thus it is not sufficient to
get this line with grep; one needs to eval it and output the result.

GNU Autoconf 2.69 was including the whole $ac_cs_config value inside
double quotes, and the previous code also had issues with configure
arguments containing special characters (but spaces were fine).