Fix compare_uid() to work with large UID values.
The function was pulled from the other sort methods used in mutt.  But
those don't work properly for a 32-bit unsigned value.  If the
difference between two UID values is greater than a signed int can
represent, it will sort improperly.

Some of the other sort functions need to be fixed too, but that can be
done in master.