QRESYNC: don't increment msn for missing headers.
Commit 74ce032f moved the msn increment outside the check for a
missing header.  The idea was that a header-cache hole needed to be
accounted for.  However, a header can also be missing just by a VANISH
happening the last time the mailbox was open.

Since the header won't be in the uid_hash, cmd_parse_vanished() can't
decrement subsequent msns, so it need to be done in this loop.

If there is a header cache hole, or another problem, verify_qresync()
will detect a problem and try again without QRESYNC.

However if there were an actual stored "blank", do increment, since
there was no UID to process a vanish on.