Implement <select-entry> as <display-message> in index menu.
The default keybindings of <display-message> overshadow those of
<select-entry>.  However, for someone who rebinds (globally or just in
the index menu), it's still possible to invoke <select-entry> inside
the index.

It's debatable if <display-message> is the only choice in this case.
For instance, <attach-message> in the compose menu could conceivably
make use of the function.  However that opens another set of issues,
and as someone on mutt-users said is solving a problem no one has
complained about.

For now, I think <display-message> is the best choice outside of
displaying an error message, so make it do that.