Don't use subtraction for qsort numeric value comparisons.
Subtraction can overflow, resulting in incorrect sorts.  This is
especially a concern for the date and size sorting, whose fields are
greater than an 'int' size.

The index values should be okay, but it's better to be consistent and
avoid any possible issues.

Define a macro, mutt_numeric_cmp(), that uses direct comparison rather
than subtraction.