Improve smime_verify_one() error handling and general flow.
mutt_decode_attachment() does not change offset, length, or type.  The
previous commit added a const modifier to its BODY parameter to assert
that point.

The save/restore appears to be copy/pasted from other code.  The
length/type are assigned to if the BODY is passed to a helper function
after the decode, but that's not the case here.  Remove the
backup/restore code of the sigbdy fields.

The smime_invoke_verify() call uses the signedfile directly, not the
STATE parameter.  Remove the code to swap s->fpin before calling
smime_invoke_verify().

Improve the error handling to properly close file handles and unlink
all temp files.