Add $attach_save_charset_convert quadoption.
This prompts to allow charset conversion of received text-type
attachments.

Mutt automatically converts when sending to a value from
$send_charset.  If an attachment were automatically converted from
utf-8 to iso-8859-1 on the sending side, then on the receiving
side (also 99.9% likely to be utf-8 nowadays) it may not be helpful to
save the the originally utf-8 attachment as iso-8859-1.  This can be
especially destructive for attached patches or translation updates.

Add the quadoption, defaulting to ask-yes so people know about it.
With a quadoption, it can be easily completely turned off or made the
default if the user desires.

Add code to skip the prompt if converting from ascii to extended
ascii (e.g. utf-8 or iso-8859-*).  There may be others too but this
should cover almost all cases nowadays.