Ignore $delete when sync'ing with $maildir_trash set.
When closing a mailbox, the $delete quadoption is not consulted to
determine whether to write out the T flags.  However, for some reason,
sync was doing so.

Since we aren't "purging" the messages, the prompt doesn't make sense.
Make close and sync consistent by skipping the $delete check for sync.