Don't force SMTP authentication without a username or client cert.
See Debian bug 1001013.  This is a regression from commit 191b0513,
which fixed SMTP client certification authentication (see GitLab
ticket 336).  However, that commit assumed the server wouldn't
advertise AUTH if it wasn't needed, which the Debian ticket shows is
not always the case.

Client certificate support still relies on AUTH properly being set.

Add a check for either the username or a cert before calling
smtp_auth().  The username check was removed by the above mentioned
commit, but more importantly the username is no longer set by the
OpenSSL/GnuTLS client cert code.  (Which was the main problem
addressed by the commit).  So there shouldn't be an issue with adding
the username check back.