Add pager REDRAW_FLOW redraw flag when popping a menu.
A setting change or window resize could take place in the called menu,
which will require reflowing the pager when returning to it.

When converting to the menu stack, I left in manual REDRAW_FULL redraw
assignments, to be cautious and because it wasn't causing an problems.
However, with the REDRAW_FLOW added, the pager menu needs to have
those assignments removed after a menu call.