Add very basic cd path completion.
This uses existing code to partially complete, and then launch the
browser for selection.

Unfortunately, completing a path argument brings up other issues, such
as quoting, which this currently doesn't handle.  It also might be
better to rotate through completions rather than launching the file
brower.

Add a "select directory" mode to the browser, modeled on the IMAP
"enter vs select a mailbox with children" behavior: using <view-file>
to select the directory and <select-entry> to enter the directory.