Localize $reply_regexp.
This will allow translators to add additional reply prefixes common to
their locale.

Add more documentation to the option, to explain the parts of the
regular expression.  Give an example of adding more prefixes.

Note that the result of this is stored in the header cache.  Probably
people rarely were affected by this before, but with the new
localization, people may experiment and need to be aware to turn the
header cache off while testing.