Don't queue IMAP close commands.
Commit e7df4d5a changed the close commands to queue up, on the theory
that it will always be followed by a SELECT or a LOGOUT.

However, performing a <recall-message> while in the inbox will open a
*new* connection.  mutt_conn_find() prepends the new conection to the
Connections list.  The next mailbox opened will use that *new*
connection.

The effect will be the queued CLOSE won't occur before that mailbox is
opened, and so "\Deleted" messages won't be expunged on the server
right away.

If the user had $move set and reopened their $spoolfile, they would
find the read messages marked for delete but not yet expunged from the
server.